chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/cascade-of-treeboundary-crossing-rules-expected.txt

Test for casacde of treeboundary crossing rules.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL borderColorOf(getNodeInComposedTree("target")) should be rgb(0, 128, 0). Was rgb(0, 0, 255).
PASS borderColorOf(getNodeInComposedTree("host/target")) is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
